applications/core_application/pyproject.toml
[tool.poetry]
name = "core-application"
version = "0.1.0"
description = ""
authors = ["osieltorres <osieltorresdev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-configurations = "^2.5.1"
djangorestframework = "^3.15.2"
celery = "^5.4.0"
django-redis = "^5.4.0"
flower = "^2.0.1"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.2.0"
django-celery-results = "^2.5.1"
drf-spectacular = "^0.27.2"
django-filter = "^24.3"
black = "^24.8.0"
django-cors-headers = "^4.4.0"
gunicorn = "^23.0.0"
django = "5.0.6"
django-celery-beat = "^2.6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-django = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
